Add fmg as a short name for the foremerge command
`foremerge status` is a lot to type many times a day. `fmg` is the same
program under a shorter name.

The name was chosen by checking rather than guessing. `fm` is the obvious
candidate but a Go terminal file manager already installs a binary called
`fm`, and its audience overlaps ours heavily, so taking that name would
shadow a tool people installed deliberately. `fmg` is unclaimed on crates.io,
Homebrew, and Debian, and the npm and PyPI packages holding the name are
libraries that install no executable.

Rather than compile main.rs into two binary targets, which makes cargo warn
about one file in multiple targets and builds the CLI twice, the entry point
moves to `foremerge::cli::run` and both binaries are three lines over it. They
cannot drift apart, and the CLI compiles once. Clap already takes the usage
line from argv[0], so `fmg --help` names `fmg` with no extra code; `--version`
still reports the product name, which is the behaviour you want from a version
string.

Moving the CLI into the library surfaced a latent rustdoc failure: cargo skips
documenting a bin whose name collides with the lib, so an unescaped `<id>` in
a doc comment had never been linted. Fixed here.

`make verify` and `make msrv` both pass.
